+++ packages/nns/src/types/common.ts
@@ -1,9 +1,7 @@
 export type CanisterIdString = string;
 export type NeuronId = bigint;
-export type AccountIdentifier = string;
 export type E8s = bigint;
 export type Memo = bigint;
 export type PrincipalString = string;
-export type SubAccount = Uint8Array;
 
 export type Option<T> = T | undefined;